+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 1.11.0
+* `NEW` `Lua.runtime.plugin`
 * `CHG` completion: improve performance of workspace words
 * `FIX` hover: tail comments may be cutted
 * `FIX` runtime errors